[travis][appveyor] added webhook notifications
the idea is to automate bundle-repo updates. webhooks running on jancy.org
will handle notifications from appveyor and travis, and once both appveyor and
travis builds succeed for a particular commit, a script will pull a bundle repo,
update submodule with the latest commits on master branches, create a new commit
with `git submodule overview` as the commit message and push it back.